:sparkles: create audio component
