+++ src/lib/TranslationResult.svelte
@@ -1,11 +1,9 @@
 <script lang="ts">
     import { COLORS } from "./constants/colors";
     import Tags from "./Tags.svelte";
+    import Speaker from "./Speaker.svelte";
     import { goto } from '$app/navigation';
-    import { textToSpeech } from "./helpers/translate";
-
-    import speaker from '$lib/assets/speaker.png';
-    import speakerFill from '$lib/assets/speakerFill.png';
+   
     import chevron from '$lib/assets/chevron.svg';
     import { resultData } from "../stores/translateStore";
 
@@ -18,30 +16,6 @@
     export let examples: {};
     export let originalLanguage: string;
     export let translateLanguage: string;
-
-    let isAudioPlaying = false;
-
-    let speechOutput = '';
-
-
-    const handleSpeak = async() => {
-        isAudioPlaying = true;
-        if (speechOutput !==''){
-            let audio = new Audio(speechOutput);
-            await audio.play();
-        }
-        else{
-            try {
-                speechOutput = await textToSpeech(word);
-            } catch (error) {
-                console.error('Error:', error);
-                speechOutput = 'Error occurred during text-to-speech conversion.';
-            }
-        }
-
-        setTimeout(() => {isAudioPlaying = false}, 1000)
-    }
-
 
     const handleClick = () => {
         resultData.set({
@@ -67,18 +41,7 @@
             <div class="title-container">
                 <div class="title">
                     <h2 class="result">{word}</h2>
-                    <button on:click={handleSpeak}>
-                        <div class="speaker-container">
-                            {#if isAudioPlaying}
-                            <img class="speaker" alt="speaker icon" src={speakerFill} />
-                            {:else}
-                            <img class="speaker" alt="speaker icon" src={speaker} />
-                            {/if}
-                        </div>
-                    </button>
-                    {#if speechOutput!== ''}
-                        <audio autoplay><source type="audio/mpeg" src={speechOutput}></audio>
-                    {/if}
+                    <Speaker phrase={word} />
                 </div>
                 <div class="subtitle">
                     <i class="subtitle-text phonetic">{phoneticSpelling}</i>
@@ -108,17 +71,6 @@
 
 
 <style>
-    button{
-        border: 0;
-        padding: 0;
-        margin: 0;
-        background-color: transparent;
-    }
-
-    button:hover{
-        cursor: pointer;
-    }
-
     .container{
         max-width: 21.375rem;
         min-width: 16rem;
@@ -241,17 +193,6 @@
         font-style: normal;
         font-weight: 700;
         line-height: normal;
-    }
-
-    .speaker{
-        height: 30px;
-        padding: 7px;
-    }
-
-    .speaker-container {
-        max-height: 1.5rem;
-        display: flex;
-        align-items: center;
     }
 
     .subtitle{
