+++ src/lib/helpers/helperFunctions.ts
